refactor(codex): split app-server attempt helpers
